requirements.txt
streamlit
torch
torchvision
numpy
pillow
scikit-image
fastai
huggingface-hub